Add interview samples, Coderbyte challenges, and expanded Q&A docs.
Includes BigDecimal employee salary helpers, longest-substring variants, Spring Boot proxy/TX docs, Mphasis architect notes, and Coderbyte Java solutions.

+++ src/main/java/com/boot2/Employee.java
@@ -1,19 +1,25 @@
 package com.boot2;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import lombok.AllArgsConstructor;
 
+import java.math.BigDecimal;
+
+/**
+ * Simple employee model. {@code salary} is {@link BigDecimal} so money is exact
+ * (prefer over {@code double}/{@code float}).
+ */
 @Getter
 @Setter
 @ToString
 @AllArgsConstructor
 public class Employee {
 
-   private String firstName;
-   private String lastName;
-   private int deptId;
-   private double salary;
+    private String firstName;
+    private String lastName;
+    private int deptId;
+    private BigDecimal salary;
 
 }
